Fix warning: implicit conversion changes signedness
visp/modules/io/src/image/private/vpImageIoOpenCV.cpp: In function ‘void writePNGtoMemOpenCV(const vpImage<vpRGBa>&, std::vector<unsigned char>&, bool)’:
visp/modules/io/src/image/private/vpImageIoOpenCV.cpp:306:47: warning: conversion to ‘std::size_t’ {aka ‘long unsigned int’} from ‘sizetype’ may change the sign of the result [-Wsign-conversion]
  306 |     unsigned char *bitmap = new unsigned char[static_cast<size_t>(height * width * channels)];
      |                                               ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

+++ modules/io/src/image/private/vpImageIoOpenCV.cpp
@@ -303,7 +303,7 @@ void writePNGtoMemOpenCV(const vpImage<vpRGBa> &I_color, std::vector<unsigned ch
     }
   }
   else {
-    unsigned char *bitmap = new unsigned char[static_cast<size_t>(height * width * channels)];
+    unsigned char *bitmap = new unsigned char[static_cast<unsigned int>(height * width * channels)];
     vpImageConvert::RGBaToRGB(reinterpret_cast<unsigned char *>(I_color.bitmap), bitmap, static_cast<unsigned int>(height*width));
 
     cv::Mat3b img(static_cast<int>(height), static_cast<int>(width), reinterpret_cast<cv::Vec3b *>(bitmap));
